delete confirm added

+++ src/pages/dashboard/MyOrders2.js
@@ -2,12 +2,9 @@ import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { Link } from "react-router-dom";
-import Swal from "sweetalert2";
+import { toast } from "react-toastify";
 import auth from "../../firebase.init";
-import Delete from "../../shared/svgIcon/Delete";
-import Pay from "../../shared/svgIcon/Pay";
-// import deleteIcon from "../../images/icons/delete-bin-4-line.svg";
-// import payicon from '../../images/icons/bank-card-line.svg'
+import deleteIcon from "../../images/icons/delete-bin-4-line.svg";
 
 const MyOrders = () => {
   const [orders, setOrders] = useState([]);
@@ -22,47 +19,27 @@ const MyOrders = () => {
       });
   }, [user?.email]);
 
-  const payNow = (id) =>{
-    console.log('pay now');
-  }
-
   const deleteOrder = (id) => {
-    Swal.fire({
-      title: 'Are you sure?',
-      text: "You won't be able to revert this!",
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonColor: '#3085d6',
-      cancelButtonColor: '#d33',
-      confirmButtonText: 'Yes, delete it!'
-    }).then((result) => {
-      if (result.isConfirmed) {
-
-// main delete function 
-  axios
-    .delete(`https://manufacturer-website-ks.herokuapp.com/order/${id}`, {
-      method: "delete",
-    })
-    .then((res) => {
-      if (res.data.deletedCount) {
-        axios
-          .get(
-            `https://manufacturer-website-ks.herokuapp.com/my-orders/${user?.email}`
-          )
-          .then((res) => {
-            setOrders(res.data);
-          });
-          Swal.fire(
-            'Deleted!',
-            'The order has been deleted.',
-            'success'
-          )
-      }
-    });
+    const proceed = window.confirm("Are you sure to delete this item?");
+    if (proceed) {
+      axios
+        .delete(`https://manufacturer-website-ks.herokuapp.com/order/${id}`, {
+          method: "delete",
+        })
+        .then((res) => {
+          if (res.data.deletedCount) {
+            toast.warn("Order Deleted");
+            axios
+              .get(
+                `https://manufacturer-website-ks.herokuapp.com/my-orders/${user?.email}`
+              )
+              .then((res) => {
+                setOrders(res.data);
+              });
+          }
+        });
     }
-   })
-   
-};
+  };
 
   return (
     <div>
@@ -117,22 +94,13 @@ const MyOrders = () => {
                 <td>${order.total_price}</td>
                 <td>{order.status}</td>
                 <td>
-                 <div className="flex gap-2 items-center justify-start">
-                 <button
-                    htmlFor="delete-modal" data-tip="Pay Now"
-                    className="w-8 h-8 tooltip flex"
-                    onClick={() => payNow(order._id)}
-                  >
-                   <Pay/>
-                  </button>
-                 <button
-                    htmlFor="delete-modal" data-tip="Cancel Order"
-                    className="w-8 h-8 tooltip flex"
+                  <button
+                    htmlFor="delete-modal"
+                    className="w-10 h-10 flex items-center justify-center rounded-full bg-base-200"
                     onClick={() => deleteOrder(order._id)}
                   >
-                   <Delete/>
+                    <img src={deleteIcon} alt="" />
                   </button>
-                 </div>
                 </td>
               </tr>
             ))}
